add button simulation functions
We save the pid of the "owner" process as persistent term, which is
then used to send button press events to the owner.

The button functions are very simplistic, and are there for simple
interactive testing. The function arguments are not checked for type
or range. The sole purpose of this module is to enable us run the
updated simulator.